[DOP-1039] - add dev release
