+++ .github/workflows/dev-release.yml
@@ -1,0 +1,90 @@
+name: Dev release
+
+on:
+  push:
+    branches-ignore:
+    - dependabot/**
+    - pre-commit-ci-update-config
+    - master
+  workflow_dispatch:
+
+jobs:
+  wait-tests:
+    name: Wait for tests
+    runs-on: ubuntu-latest
+
+    steps:
+    - uses: fountainhead/action-wait-for-check@v1.1.0
+      id: wait-for-tests
+      with:
+        token: ${{ secrets.GITHUB_TOKEN }}
+        checkName: Tests done
+        ref: ${{ github.sha }}
+        timeoutSeconds: 3600
+
+    - name: Fail the Build
+      uses: cutenode/action-always-fail@v1
+      if: steps.wait-for-tests.outputs.conclusion != 'success'
+
+  wait-codeql:
+    name: Wait for CodeQL
+    runs-on: ubuntu-latest
+
+    steps:
+    - uses: fountainhead/action-wait-for-check@v1.1.0
+      id: wait-for-codeql
+      with:
+        token: ${{ secrets.GITHUB_TOKEN }}
+        checkName: Analyze
+        ref: ${{ github.sha }}
+        timeoutSeconds: 3600
+
+    - name: Fail the Build
+      uses: cutenode/action-always-fail@v1
+      if: steps.wait-for-codeql.outputs.conclusion != 'success'
+
+  release:
+    name: Release package
+    runs-on: ubuntu-latest
+    needs: [wait-tests, wait-codeql]
+
+    steps:
+    - name: Checkout code
+      uses: actions/checkout@v3
+      with:
+        fetch-depth: 0
+
+    - name: Set up Python 3.10
+      uses: actions/setup-python@v4
+      with:
+        python-version: '3.10'
+
+    - name: Cache pip
+      uses: actions/cache@v3
+      with:
+        path: ~/.cache/pip
+        key: ${{ runner.os }}-python-3.10-release-${{ hashFiles('requirements*.txt') }}
+        restore-keys: |
+          ${{ runner.os }}-python-3.10-release-${{ hashFiles('requirements*.txt') }}
+          ${{ runner.os }}-python-3.10-tests-${{ hashFiles('requirements*.txt') }}
+          ${{ runner.os }}-python-3.10-codeql-${{ hashFiles('requirements*.txt') }}
+          ${{ runner.os }}-python-3.10-release-
+          ${{ runner.os }}-python-3.10-tests-
+          ${{ runner.os }}-python-3.10-codeql-
+          ${{ runner.os }}-python
+          ${{ runner.os }}-
+
+    - name: Upgrade pip
+      run: python -m pip install --upgrade pip setuptools wheel
+
+    - name: Install dependencies
+      run: pip install -r requirements.txt
+
+    - name: Build package
+      run: python setup.py sdist bdist_wheel
+
+    - name: Publish package
+      uses: pypa/gh-action-pypi-publish@release/v1
+      with:
+        password: ${{ secrets.TEST_PYPI_API_TOKEN }}
+        repository_url: https://test.pypi.org/legacy/
